pid がない ( 初回など ) はスルー

+++ files/kill_unicorn.sh
@@ -1,1 +1,3 @@
-kill `cat ~/$1/tmp/pids/unicorn.pid`
+if [ -e ~/$1/tmp/pids/unicorn.pid ]; then
+  kill `cat ~/$1/tmp/pids/unicorn.pid`
+fi
